PermissionGrant Signing (#746)
This is the first in a series of PRs that enable users to use permission grants via the web5 agent.

In a subsequent PR these features will be used within `@web5/api` to fetch and cache the grants, and use them accordingly. The same will happen for the `sync manager`. Each higher-level abstraction that calls `processMessage` will manage it's own fetching/caching and using of the grants themselves.

Within this PR:

- a `granteeDid` optional property has been added to `ProcessDwnRequest`  representing the DID that is signing the message and has been granted a permission.
- a `signAsOwnerDelegate` optional property has been added to `ProcessDwnRequest` showing intent for signing as an owner using a delegate grant.

Some temporary helper methods have been added to the `dwn` api that will eventually move to it's own `permissions` api in a subsequent PR. These include:
- `fetchGrants`
- `isGrantRevoked`
- `createGrant`
- `createRevocation`

A utility method `matchGrantFromArray` has been added to the general utilities, this may be moved to a static method within the `permissions` api in a subsequent PR.