Conflicting Sync error. (#857)
When I added the one-shot `sync()` method I put in a guard to prevent it from being called while an interval was running. This caused a bug in the interval, so I've modified the code to account for it.

Also updated some type docs to get rid of the warning.