add grant to Record class methods (#859)
This PR fixes a bug where the `Record` class method which is retrieved with queries, reads, and writes did not respect the delegated permission state. The `Record` class now accepts an optional PermissionsApi so that it can share the TTL cache with a caller. If one is not provided a new one is instantiated from the agent that is passed.

Took advantage of removing the specific `CachedPermissions` class and wrapped it into the `PermissionsApi` interface.

Refactored some unnecessary abstractions when fetching the grant, added tests for missing cases and subscription cases.