Ensure `protocolRole` is maintained between query/read and subscribe/read. (#954)
Before this PR that were some inconsistencies with using `protocolRole`.

There were instances where a user would query using a role but not be able to read the data of the given record because the role was not being applied. Same would happen during update/delete.

This PR allows the READ operation to inherit the `protocolRole` used for a `query` or `subscribe` if it exists.
Additionally it provides the user the ability to provide a different role when performing an `update` or `delete` operation.