records can be updated without being stored (#957)
This PR resolves the issue where a user is forced to `store()` a record prior to updating it. Now a user can pass `store: false` to the update options to signal they do not wish to store the updated record.

The default behavior remains the same, effectively store is defaulted to true. If the initial record was not already stored, it will throw with an error as it does today forcing the user to explicitly `store()` the record prior to updating it unless they do not want to store the udpate.

Resolves https://github.com/TBD54566975/web5-js/issues/955